--- src/main/java/com/waggle/apis/auth/dto/GoogleUserInfo.java
@@ -1,31 +1,0 @@
-package com.waggle.apis.auth.dto;
-
-import lombok.AllArgsConstructor;
-
-import java.util.Map;
-
-@AllArgsConstructor
-public class GoogleUserInfo implements OAuth2UserInfo {
-
-    private Map<String, Object> attributes;
-
-    @Override
-    public String getProviderId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getProvider() {
-        return "google";
-    }
-
-    @Override
-    public String getName() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) attributes.get("email");
-    }
-}
