--- src/main/java/com/waggle/apis/auth/dto/KakaoUserInfo.java
@@ -1,32 +1,0 @@
-package com.waggle.apis.auth.dto;
-
-import lombok.AllArgsConstructor;
-
-import java.util.Map;
-
-@AllArgsConstructor
-public class KakaoUserInfo implements OAuth2UserInfo{
-
-    private Map<String, Object> attributes;
-
-    @Override
-    public String getProviderId() {
-        return String.valueOf(attributes.get("id"));
-    }
-
-    @Override
-    public String getProvider() {
-        return "kakao";
-    }
-
-    @Override
-    public String getName() {
-        Map<String, Object> profile = (Map<String, Object>) ((Map<String, Object>) attributes.get("kakao_account")).get("profile");
-        return (String) profile.get("nickname");
-    }
-
-    @Override
-    public String getEmail() {
-        return (String) ((Map<String, Object>) attributes.get("kakao_account")).get("email");
-    }
-}
