--- src/main/java/com/waggle/apis/auth/dto/OAuth2UserInfo.java
@@ -1,8 +1,0 @@
-package com.waggle.apis.auth.dto;
-
-public interface OAuth2UserInfo {
-    String getProviderId();
-    String getProvider();
-    String getName();
-    String getEmail();
-}
