--- src/main/java/com/waggle/apis/auth/entity/RefreshToken.java
@@ -1,27 +1,0 @@
-package com.waggle.apis.auth.entity;
-
-import com.waggle.apis.user.entity.User;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.UUID;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-@Getter
-@Table(name = "refresh_tokens")
-public class RefreshToken {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id")
-    private UUID id;
-
-    @OneToOne
-    @JoinColumn(name = "user_id", referencedColumnName = "id", unique = true, nullable = false)
-    private User user;
-
-    @Column(name = "token", nullable = false)
-    private String token;
-}
