--- src/main/java/com/waggle/apis/user/entity/User.java
@@ -1,42 +1,0 @@
-package com.waggle.apis.user.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-@Builder
-@Getter
-@Table(name = "users")
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "id")
-    private UUID id;
-
-    @Column(name = "name", nullable = false, length = 5)
-    private String name;
-
-    @Column(name = "email", nullable = false, length = 50)
-    private String email;
-
-    @Column(name = "provider", nullable = false, length = 10)
-    private String provider;
-
-    @Column(name = "provider_id", nullable = false, length = 50)
-    private String providerId;
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false, length = 20)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at", length = 20)
-    private LocalDateTime updatedAt;
-}
