--- src/main/java/com/waggle/apis/user/entity/UserRepository.java
@@ -1,16 +1,0 @@
-package com.waggle.apis.user.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.UUID;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,UUID> {
-    @Query("SELECT u FROM User u WHERE u.id = :userId")
-    Optional<User> findByUserId(UUID userId);
-
-    User findByProviderId(String providerId);
-}
