--- src/main/java/com/waggle/auth/OAuth2LoginFailureHandler.java
@@ -1,24 +1,0 @@
-package com.waggle.auth;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class OAuth2LoginFailureHandler extends SimpleUrlAuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response, AuthenticationException exception) throws IOException, ServletException {
-        log.error("LOGIN FAILED : {}", exception.getMessage());
-        super.onAuthenticationFailure(request, response, exception);
-    }
-}
