+++ src/main/java/com/waggle/domain/auth/dto/AccessTokenResponse.java
@@ -1,0 +1,12 @@
+package com.waggle.domain.auth.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class AccessTokenResponse {
+    @JsonProperty("access_token")
+    private String accessToken;
+}
