+++ src/main/java/com/waggle/domain/auth/service/AuthService.java
@@ -1,0 +1,9 @@
+package com.waggle.domain.auth.service;
+
+import com.waggle.domain.auth.dto.AccessTokenResponse;
+import com.waggle.domain.user.entity.User;
+
+public interface AuthService {
+    AccessTokenResponse reissueAccessToken(String refreshToken);
+    User getCurrentUser();
+}
