--- src/main/java/com/waggle/global/ApiResponse.java
@@ -1,29 +1,0 @@
-package com.waggle.global;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.waggle.global.interfaces.Status;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-
-@Getter
-@RequiredArgsConstructor
-public class ApiResponse<T> {
-    @JsonProperty("isSuccess")
-    private final Boolean isSuccess;
-    private final int code;
-    private final String message;
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    private final T payload;
-    public static <T> ResponseEntity<ApiResponse<T>> onSuccess(Status code, T data) {
-        ApiResponse<T> response = new ApiResponse<>(true, code.getReasonHttpStatus().getCode(), code.getReasonHttpStatus().getMessage(), data);
-        return ResponseEntity.status(code.getReasonHttpStatus().getHttpStatus()).body(response);
-    }
-
-    public static <T> ResponseEntity<ApiResponse<T>> onFailure(Status code) {
-        ApiResponse<T> response = new ApiResponse<>(false, code.getReasonHttpStatus().getCode(), code.getReasonHttpStatus().getMessage(), null);
-        return ResponseEntity.status(code.getReasonHttpStatus().getHttpStatus()).body(response);
-    }
-}
