+++ src/main/java/com/waggle/global/exception/JwtTokenException.java
@@ -1,0 +1,17 @@
+package com.waggle.global.exception;
+
+import com.waggle.global.response.code.JwtTokenErrorResponseCode;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class JwtTokenException extends RuntimeException {
+    private final JwtTokenErrorResponseCode jwtTokenErrorStatus;
+
+    @Override
+    public String getMessage() {
+        return jwtTokenErrorStatus.getMessage();
+    }
+}
