--- src/main/java/com/waggle/global/interfaces/Status.java
@@ -1,8 +1,0 @@
-package com.waggle.global.interfaces;
-
-import com.waggle.global.dto.ResponseDto;
-
-public interface Status {
-    public ResponseDto getReason();
-    public ResponseDto getReasonHttpStatus();
-}
