+++ src/main/java/com/waggle/global/response/code/ResponseCode.java
@@ -1,0 +1,8 @@
+package com.waggle.global.response.code;
+
+import com.waggle.global.response.ApiResponseMessage;
+
+public interface ResponseCode {
+    public ApiResponseMessage getReason();
+    public ApiResponseMessage getReasonHttpStatus();
+}
