+++ src/main/java/com/waggle/global/secure/oauth2/adapter/GoogleUserInfoAdapter.java
@@ -1,0 +1,31 @@
+package com.waggle.global.secure.oauth2.adapter;
+
+import lombok.AllArgsConstructor;
+
+import java.util.Map;
+
+@AllArgsConstructor
+public class GoogleUserInfoAdapter implements OAuth2UserInfo {
+
+    private Map<String, Object> attributes;
+
+    @Override
+    public String getProviderId() {
+        return (String) attributes.get("sub");
+    }
+
+    @Override
+    public String getProvider() {
+        return "google";
+    }
+
+    @Override
+    public String getName() {
+        return (String) attributes.get("name");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) attributes.get("email");
+    }
+}
