+++ src/main/java/com/waggle/global/secure/oauth2/adapter/KakaoUserInfoAdapter.java
@@ -1,0 +1,32 @@
+package com.waggle.global.secure.oauth2.adapter;
+
+import lombok.AllArgsConstructor;
+
+import java.util.Map;
+
+@AllArgsConstructor
+public class KakaoUserInfoAdapter implements OAuth2UserInfo{
+
+    private Map<String, Object> attributes;
+
+    @Override
+    public String getProviderId() {
+        return String.valueOf(attributes.get("id"));
+    }
+
+    @Override
+    public String getProvider() {
+        return "kakao";
+    }
+
+    @Override
+    public String getName() {
+        Map<String, Object> profile = (Map<String, Object>) ((Map<String, Object>) attributes.get("kakao_account")).get("profile");
+        return (String) profile.get("nickname");
+    }
+
+    @Override
+    public String getEmail() {
+        return (String) ((Map<String, Object>) attributes.get("kakao_account")).get("email");
+    }
+}
