+++ src/main/java/com/waggle/global/secure/oauth2/adapter/OAuth2UserInfo.java
@@ -1,0 +1,8 @@
+package com.waggle.global.secure.oauth2.adapter;
+
+public interface OAuth2UserInfo {
+    String getProviderId();
+    String getProvider();
+    String getName();
+    String getEmail();
+}
